+++ master-nafsdm/nafsdmctl/__main__.py
@@ -57,7 +57,7 @@ elif (sys.argv[1] == "remove"):
 
             wasRemoved = False
             for currentLine in rawDomains.split("\n"):
-                if len(currentLine.split()) == 4:
+                if len(currentLine.split()) == 5:
                     if sys.argv[2] in currentLine:
                         wasRemoved = True
                     else:
@@ -88,6 +88,74 @@ elif (sys.argv[1] == "list"):
         print(currentLine)
 
     print(longLine)
+elif (sys.argv[1] == "edit"):
+    if (len(sys.argv) == 3):
+        if "." in sys.argv[2]:
+            f = open("/home/master-nafsdm/data/domains.txt")
+            rawDomains = f.read()
+            f.close()
+
+            # remove config
+            os.remove("/home/master-nafsdm/data/domains.txt")
+
+            wasEdited = False
+            for currentLine in rawDomains.split("\n"):
+                if len(currentLine.split()) == 5:
+                    if sys.argv[2] in currentLine:
+
+                        # domain name
+                        domain_name = raw_input("Enter new domain name (blank for no change): ")
+                        if domain_name == "":
+                            domain_name = currentLine.split()[0]
+
+                        # master IP
+                        master_ip = raw_input("Enter new master IP (blank for no change): ")
+                        if master_ip == "":
+                            master_ip = currentLine.split()[1]
+
+                        # comment
+                        comment = raw_input("Enter new comment (blank for no change): ")
+                        if comment == "":
+                            comment = currentLine.split()[2]
+
+                        # slaves
+                        slaves = raw_input("Enter new slaves (seperated with '.') (blank for no change): ")
+                        if "." in slaves:
+                            if slaves == "":
+                                slaves = currentLine.split()[3]
+                        else:
+                            print("syntax error: invalid slaves. Continuing anyways (set to same as before)")
+                            slaves = currentLine.split()[3]
+
+                        # DNSSEC
+                        dnssec = raw_input("Edit DNSSEC status (only type yes or no) (blank for no change): ")
+                        if dnssec == "yes":
+                            dnssec = "dnssec.yes"
+                        elif dnssec == "no":
+                            dnssec = "dnssec.no"
+                        else:
+                            dnssec = currentLine.split()[4]
+                            print("syntax error: only yes or no supported. Value set to same as before.")
+
+                        writeLine = domain_name + " " + master_ip + " " + comment + " " + slaves + " " + dnssec + "\n"
+                        f = open("/home/master-nafsdm/data/domains.txt", "a")
+                        f.write(writeLine)
+                        f.close()
+
+                        wasEdited = True
+                    else:
+                        f = open("/home/master-nafsdm/data/domains.txt", "a")
+                        f.write(currentLine + "\n")
+                        f.close()
+
+            if wasEdited == True:
+                print("nafsdmctl: edit succesful")
+            if wasEdited == False:
+                print("nafsdmctl: edit failed - invalid domain name?")
+        else:
+            print("syntax error: invalid domain name?")
+    else:
+        print("syntax error: 'nafsdmctl edit domain.tld' is correct syntax")
 else:
     # just prints some of the syntaxes and exists as an error
     print("syntax error: please use correct argument." + "\n" +