+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Version 1.2 (Stauts: DEVELOPMENT)
 * Added DNSSEC support.
 * Planned: Added upgrade scripts.
 * Planned: Added support for automatic upgrades (1.12 and forwards)
-* Planned: New edit function for nafsdmctl (master).
+* New edit function for nafsdmctl (master).
 
 Version 1.1 (Status: RELEASED)
 * New version files.